+++ packages/volto-separator-block/src/components/Separator/Edit.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Edit = ({ data }) => {
-  return <div />;
-};
-
-export default Edit;