+++ src/components/ButtonBlock/ButtonSidebar.jsx
@@ -1,8 +1,0 @@
-import React from 'react';
-import ButtonData from './ButtonData';
-
-const ButtonSidebar = (props) => {
-  return <ButtonData {...props} />;
-};
-
-export default ButtonSidebar;